Implementamos los iconos del menu asi como la clase encargada de recargar la base de dato con los servicios web del ayuntamiento.
