+++ app/src/main/java/unican/es/grupo4_tus_santander/View/Main/ListFuncionesMainAdapter.java
@@ -1,12 +1,14 @@
-package unican.es.grupo4_tus_santander.View;
+package unican.es.grupo4_tus_santander.View.Main;
 
 import android.content.Context;
+import android.media.Image;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import unican.es.grupo4_tus_santander.R;
@@ -27,9 +29,34 @@ public class ListFuncionesMainAdapter extends ArrayAdapter {
     public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
         LayoutInflater layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
         View viewRow = layoutInflater.inflate(R.layout.custom_list_funciones_main_layout,null,true);
-        TextView textViewNameFuncion = (TextView) viewRow.findViewById(R.id.textViewNameFuncion);
 
-        textViewNameFuncion.setText( funcionesMain[position]);
+        TextView textViewNameFuncion = (TextView) viewRow.findViewById(R.id.textViewNameFuncion);
+        ImageView imagenViewFuncion = (ImageView) viewRow.findViewById(R.id.ImagenFuncion);
+
+        textViewNameFuncion.setText(funcionesMain[position]);
+
+        switch(position){
+            case 0:
+                imagenViewFuncion.setImageResource(R.drawable.ic_favoritos);
+                break;
+            case 1:
+                imagenViewFuncion.setImageResource(R.drawable.ic_lineas);
+                break;
+            case 2:
+                imagenViewFuncion.setImageResource(R.drawable.ic_paradas);
+                break;
+            case 3:
+                imagenViewFuncion.setImageResource(R.drawable.ic_tarifas);
+                break;
+            case 4:
+                imagenViewFuncion.setImageResource(R.drawable.ic_restricciones);
+                break;
+            case 5:
+                imagenViewFuncion.setImageResource(R.drawable.ic_ajustes);
+                break;
+            default:
+                break;
+        }
 
         return viewRow;
     }
