+++ app/src/main/res/drawable/ic_restricciones.xml
@@ -1,0 +1,4 @@
+<vector android:height="40dp" android:viewportHeight="24.0"
+    android:viewportWidth="24.0" android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M1,21h22L12,2 1,21zM13,18h-2v-2h2v2zM13,14h-2v-4h2v4z"/>
+</vector>
