+++ app/src/main/res/layout/activity_main.xml
@@ -4,6 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical" >
+
     <ProgressBar
         android:id="@+id/progressBar"
         style="?android:attr/progressBarStyleHorizontal"
@@ -14,7 +15,8 @@
         android:backgroundTint="@color/colorPrimary"
         android:layout_below="@+id/framelauout"
         android:indeterminateTint="#1a09d6"
-        android:layout_marginTop="-7dp"/>
+        android:layout_marginTop="-7dp"
+        android:visibility="invisible"/>
 
 
     <ListView
